# Restore hashmap node creation optimizatiom