Restore hashmap node creation optimization from #76 that got lost in the latest refactorings

Before

```
test hashmap_insert_mut_100    ... bench:       6,938.23 ns/iter (+/- 113.80)
test hashmap_insert_mut_1000   ... bench:      76,059.14 ns/iter (+/- 705.79)
test hashmap_insert_mut_10000  ... bench:     830,938.95 ns/iter (+/- 29,548.33)
```

After

```

test hashmap_insert_mut_100    ... bench:       5,872.88 ns/iter (+/- 72.71)
test hashmap_insert_mut_1000   ... bench:      63,136.52 ns/iter (+/- 577.20)
test hashmap_insert_mut_10000  ... bench:     741,631.21 ns/iter (+/- 120,330.27)
```